// src/enums/SocialNetworkEnum.ts
export enum SocialNetworkEnum {
    Facebook = "Facebook",
    Linkedin = "Linkedin",
    X = "X",
    Youtube = "Youtube",
    Tiktok = "Tiktok",
    Telegram = "Telegram",
    Instagram = "Instagram",
    Pinterest = "Pinterest",
    Reddit = "Reddit",
    Google = "Google",
    Bluesky = "Bluesky",
    Whatsapp = "Whatsapp",
    Wechat = "Wechat"
} 